cleaner

+++ ft_putnbr_fd.c
@@ -18,10 +18,10 @@ void	ft_putnbr_fd(int nb, int fd)
 	if (nb < 0)
 	{
 		ft_putchar_fd('-', fd);
-		nbr = (unsigned int)(nb * -1);
+		nbr = (nb * -1);
 	}
 	else
-		nbr = (unsigned int)nb;
+		nbr = nb;
 	if (nbr >= 10)
 		ft_putnbr_fd(nbr / 10, fd);
 	ft_putchar_fd((char)(nbr % 10 + '0'), fd);
